requirements-gpu.txt
numpy==1.23.4
opencv-python-headless==4.6.0.66
onnx==1.12.0
onnxruntime-gpu==1.13.1

